Palautettu 4. osan tehtävät

+++ osa4/blogilista/controllers/blogs.js
@@ -1,32 +1,56 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
+    const user = request.user
+    if (user === null) {
+        response.status(401).json({ error: 'token invalid' })
+    }
+
     const body = request.body
+    if (!body.title || !body.url) {
+        response.status(400).json({ error: "missing fields" })
+    }
 
     const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes
+        likes: body.likes,
+        user: user._id
     })
 
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-    const blog = await Blog.findByIdAndDelete(request.params.id)
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+    const user = request.user
+    if (user === null) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
     if (!blog) {
         return response.status(404).end();
     }
-    response.status(204).end()
+
+    if ( blog.user._id.toString() === user._id.toString() ) {   
+        await Blog.findByIdAndDelete(request.params.id)
+        return response.status(204).end()
+    } else {
+        response.status(401).json({ error: "You don't have permission to delete"})
+    }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
